pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "idandsso"
version = "0.1.0"
description = "Django based IDP integration and SSO adapter"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name="Eike Hinderk Jürrens", email="e.h.juerrens@52north.org" },
]
license = "GPL-3.0-or-later"
license-files = [
    "LICENSE",
    "license_header.txt",
]
dependencies = [
    "django>=5.1",
    "django-allauth[socialaccount]>=65.4",
    "loguru>=0.7.3",
    "python-keycloak==5.8.1",
]

[tool.hatch.build.targets.wheel]
include = [
    "idandsso/**/*.py",
    "idandsso/locale/**/*",
    "idandsso/static/**/*",
    "idandsso/templates/**/*.html",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.uv]
dev-dependencies = [
    "ruff>=0.11.3",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.0",
    "pytest-django>=4.11.1",
    "pre-commit>=4.5.1",
]
